# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "U1R API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get all user list",
                "operationId": "d64c69c227bcaa94189f8552f9336c55",
                "parameters": [
                    {
                        "name": "role",
                        "in": "query",
                        "description": "user role",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "user",
                                "seller",
                                "admin"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get all user info Successfully"
                    }
                }
            }
        },
        "/api/profile": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get authenticated user",
                "operationId": "0ed57baf5e815c7e71688207265bf99b",
                "responses": {
                    "200": {
                        "description": "Get user info Successfully"
                    }
                },
                "security": {
                    "bearerAuth": {}
                }
            },
            "patch": {
                "tags": [
                    "User"
                ],
                "summary": "Update User",
                "operationId": "8d657c4de8c480f5191c0271468608af",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "role"
                                ],
                                "properties": {
                                    "first_name": {},
                                    "last_name": {},
                                    "image": {},
                                    "district": {},
                                    "city": {},
                                    "postal_code": {},
                                    "address": {},
                                    "phone_no": {},
                                    "role": {},
                                    "is_active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sub Category Updated Successfully"
                    }
                }
            }
        },
        "/api/forgot-password": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Retrive forgotten password",
                "operationId": "08073bea29a30ce2dccaab066fb7e8de",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {}
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mail send Successfully"
                    }
                }
            }
        },
        "/api/reset-password": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Reset new password",
                "operationId": "468fc10e8c2b7cf5f33e3b5beed5dc00",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "token",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "email": {},
                                    "token": {},
                                    "password": {},
                                    "password_confirmation": {}
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password changed Successfully"
                    }
                }
            }
        },
        "/api/utility-requests": {
            "get": {
                "tags": [
                    "Utility Request"
                ],
                "summary": "Get Utility Request List",
                "operationId": "6270c7dd3b085468e2d827c9c7be3065",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get Utility Request successfully"
                    }
                }
            },
            "post": {
                "tags": [
                    "Utility Request"
                ],
                "summary": "Create Utility Request",
                "operationId": "a3682b09a43e89641b31cf8ee26e2243",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "name"
                                ],
                                "properties": {
                                    "type": {},
                                    "name": {}
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Utility Request created Successfully"
                    }
                }
            }
        },
        "/api/utility-requests/{id}": {
            "patch": {
                "tags": [
                    "Utility Request"
                ],
                "summary": "Update Utility Request",
                "operationId": "fae0d88f144e0247b34b1e46d092cbd5",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {}
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utility Request Updated Successfully"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}